+++ adminer/select.inc.php
@@ -349,8 +349,8 @@ if (!$columns && support("table")) {
 			echo script("qs('#table').onkeydown = editingKeydown;"); // keydown is not delegated
 			echo "<thead><tr>" . (!$group && $select
 				? ""
-				: "<td class='hover'><input type='checkbox' id='all-page' class='jsonly' title='" . lang('All rows on this page') . "'" . on('click', 'formCheck', '^check') . ">"
-					. " <a href='" . h($_GET["modify"] ? remove_from_uri("modify") : $_SERVER["REQUEST_URI"] . "&modify=1") . "'>" . lang('Modify') . "</a>");
+				: "<td class='hover check'><a href='" . h($_GET["modify"] ? remove_from_uri("modify") : $_SERVER["REQUEST_URI"] . "&modify=1") . "'>" . lang('Modify') . "</a>"
+					. " <input type='checkbox' id='all-page' class='jsonly' title='" . lang('All rows on this page') . "'" . on('click', 'formCheck', '^check') . ">");
 			$names = array();
 			$functions = array();
 			reset($select);
@@ -422,9 +422,9 @@ if (!$columns && support("table")) {
 					}
 					$unique_idf .= "&" . ($val !== null ? urlencode("where[" . bracket_escape($key) . "]") . "=" . urlencode($val === false ? "f" : $val) : "null%5B%5D=" . urlencode($key));
 				}
-				echo "<tr>" . (!$group && $select ? "" : "<td class='hover'>"
+				echo "<tr>" . (!$group && $select ? "" : "<td class='hover check'>"
+					. ($is_group || information_schema(DB) ? "" : "<a href='" . h(ME . "edit=" . urlencode($TABLE) . $unique_idf) . "' class='edit'>" . lang('edit') . "</a> ")
 					. checkbox("check[]", substr($unique_idf, 1), in_array(substr($unique_idf, 1), (array) $_POST["check"]))
-					. ($is_group || information_schema(DB) ? "" : " <a href='" . h(ME . "edit=" . urlencode($TABLE) . $unique_idf) . "' class='edit'>" . lang('edit') . "</a>")
 				);
 
 				reset($select);